+++ backend/users/serializers.py
@@ -1,12 +1,7 @@
-from dj_rest_auth.registration.serializers import RegisterSerializer
+# users/serializers.py
 from rest_framework import serializers
+from djoser.serializers import UserSerializer as BaseUserSerializer
 
-class CustomRegisterSerializer(RegisterSerializer):
-    username = None  # Ensures username is NOT required
-
-    def get_cleaned_data(self):
-        return {
-            'email': self.validated_data.get('email', ''),
-            'password1': self.validated_data.get('password1', ''),
-            'password2': self.validated_data.get('password2', ''),
-        }
+class UserSerializer(BaseUserSerializer):
+    class Meta(BaseUserSerializer.Meta):
+        fields = BaseUserSerializer.Meta.fields + ("credits",)